Update README.md
This project took around 15 hours over the course of a long weekend. It was designed to utilize Clarifai’s facial recognition API and integrate it into a React app that features a login/register page which allows user information to be synced to a database that was created with PostgreSQL. This allows for the tracking of when the user signed up, their email address, and the amount of entries they have submitted for facial recognition. For added security, all user passwords are hashed in the database. The background was created using the Particles.js library.